前端登陆页面的调整

+++ backend/src/main/java/org/hxw/backend/controller/UserController.java
@@ -11,7 +11,12 @@ import org.hxw.backend.service.UserService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 
+import java.util.ArrayList;
 import java.util.List;
+import java.util.stream.Collectors;
+
+import static org.hxw.backend.constant.UserConstant.ADMIN_ROLE;
+import static org.hxw.backend.constant.UserConstant.USER_LOGIN_STATE;
 
 /**
  * 用户接口
@@ -69,22 +74,45 @@ public class UserController {
      */
     @GetMapping("/search")
     public List<User> searchUsers(String username, HttpServletRequest request){
-        // 仅管理员可以查询
-//        Object userObj = request.getSession().getAttribute();
+        if(!isAdmin(request)){
+            return new ArrayList<>();
+        }
 
         QueryWrapper<User> queryWrapper = new QueryWrapper<>();
         if(StringUtils.isAnyBlank(username)){
             queryWrapper.like("username", username);
         }
-        return userService.list(queryWrapper);
+        List<User> userList = userService.list(queryWrapper);
+        return userList.stream().map(user -> {
+            user.setUserPassword(null);
+            return user;
+        }).collect(Collectors.toList());
     }
 
     @PostMapping("/delete")
-    public boolean deleteUser(@RequestBody long id){
+    public boolean deleteUser(@RequestBody long id, HttpServletRequest request){
+        if(!isAdmin(request)){
+            return false;
+        }
         if(id < 0){
             return false;
         }
         return userService.removeById(id);
+    }
+
+    /**
+     * 是否为管理员
+     * @param request
+     * @return
+     */
+    private boolean isAdmin(HttpServletRequest request){
+
+        Object userObj = request.getSession().getAttribute(USER_LOGIN_STATE);
+        User user = (User) userObj;
+        if(user == null || user.getUserRole() != ADMIN_ROLE){
+            return false;
+        }
+        return true;
     }
 
 }
